Remove tracking deprecated arrays

+++ contracts/src/v0.8/keystone/test/CapabilityRegistry_AddNodesTest.t.sol
@@ -5,7 +5,7 @@ import {BaseTest} from "./BaseTest.t.sol";
 import {CapabilityRegistry} from "../CapabilityRegistry.sol";
 
 contract CapabilityRegistry_AddNodesTest is BaseTest {
-  event NodeAdded(bytes32 p2pId, uint256 nodeOperatorId, bytes32 signer);
+  event NodeAdded(bytes32 p2pId, uint32 nodeOperatorId, bytes32 signer);
 
   uint32 private constant TEST_NODE_OPERATOR_ONE_ID = 1;
   uint32 private constant TEST_NODE_OPERATOR_TWO_ID = 2;
